+++ src/rest/decorators/headers.decorator.ts
@@ -1,7 +1,8 @@
 import {extend} from '@jscrpt/common';
 
 import {RESTClient} from '../common';
-import {RestHttpHeaders} from '../rest.interface';
+import {RestHttpHeaders, RestMethodMiddlewares} from '../rest.interface';
+import {HeadersMiddleware} from '../middlewares';
 
 /**
  * Set custom headers for a REST method
@@ -9,9 +10,11 @@ import {RestHttpHeaders} from '../rest.interface';
  */
 export function Headers(headersDef: {[key: string]: string})
 {
-    return function(_target: RESTClient, _propertyKey: string, descriptor: RestHttpHeaders)
+    return function(_target: RESTClient, _propertyKey: string, descriptor: RestHttpHeaders &
+                                                                           RestMethodMiddlewares)
     {
-        descriptor.headers = extend({}, headersDef, descriptor.headers);
+        descriptor.headers = extend(descriptor.headers ?? {}, headersDef);
+        descriptor.middlewareTypes.push(HeadersMiddleware);
 
         return descriptor;
     };