+++ src/rest/defaults.ts
@@ -1,7 +1,7 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware, IgnoredInterceptorsMiddleware, ResponseTransformMiddleware, AdditionalDataMiddleware, ProducesMiddleware, BodyParameterMiddleware, PathParameterMiddleware, QueryObjectParameterMiddleware, QueryParameterMiddleware, HeaderParameterMiddleware, HeadersMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
@@ -12,7 +12,7 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
     PathParameterMiddleware,
     QueryObjectParameterMiddleware,
     QueryParameterMiddleware,
-    // HeaderMethodMiddleware,
+    HeadersMiddleware,
     HeaderParameterMiddleware,
     ProducesMiddleware,
     AdditionalDataMiddleware,