+++ src/rest/middlewares/index.ts
@@ -8,4 +8,5 @@ export * from './bodyParameter.middleware';
 export * from './pathParameter.middleware';
 export * from './queryParameter.middleware';
 export * from './queryObjectParameter.middleware';
-export * from './headerParameter.middleware copy';
+export * from './headerParameter.middleware';
+export * from './headers.middleware';